fix: Bring content back on single post page
Content and Tags were added as ordable elements in the previous release.
Because the elements didn't exist before as ordable and the mechanism for keeping the customizer order is getting elements from the current control value, they disappeared at the theme update. They were never in the ordable control, so current value doesn't include them, and that's why they were disabled.

https://github.com/Codeinwp/neve/issues/307