Fix customizer load error with elementor-pro
Currently customizer fails to load when elementor-pro is installed but no "product" or "product-archive" template is present. This happens because on `inc/compatibility/woocommerce.php` L114, `NeveReactCustomize.elementor` is set to `{ hasElementorProductTemplate: "0", hasElementorShopTemplate: "0" }` as `Elementor::has_template` returns string value instead of bool or int. In JavaScript land, "0" is a truthy value, unlike php thus resulting in an error being thrown.